requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
black==22.8.0
click==8.1.3
coverage==6.4.4
Django==4.1.1
djangorestframework==3.14.0
exceptiongroup==1.0.0rc9
Faker==12.0.1
hypothesis==6.54.6
importlib-metadata==4.12.0
iniconfig==1.1.1
Markdown==3.4.1
mixer==7.2.2
mypy==0.971
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.3.0
zipp==3.8.1
